fix: path error when package manager not found
`pathlib.Path()` only supports strings, bytes and pathlike objects. Not
None. This was not checked and thus an error is thrown if the package
manager was not found.

With this fix, the system will still throw an error when your package
manager is not known, but it will throw the *correct* error about now
knowing your package manager.

This PR resolved #4