Added functions to copy files/dirs from the host to the platform machine (#25)
Implemented `copy_from_host` and `copy_to_host` in the communication
layer that allow files and directories to be copied from the host
machine to the platform machine and the other way around.
This allows users to sync changes from their machine to the remote
machine during benchmark setup.
For example, this is what I use in my benchmark:
```py
src = pathlib.Path(self._src_dir)
dst = pathlib.Path(self._bench_src_path)

folders_to_copy = [
    "src",
    "include",
    "benchmarking/cornell"
]

files_to_copy = [
    "build.sh",
    "CMakeLists.txt",
    "custom_config.h",
    "benchmarking/suzanne_on_table.glb"
]


# If you want to copy a folder, you need to append it with a leading / to make sure
# that the contents are not copied and that the folders are not copied into each other.
# Since pathlib does not support this (it auto strips trailing slashes), you have
# to add them manually.

for folder in folders_to_copy:
    self.platform.comm.copy_from_host(f"{src / folder}/", f"{dst / folder}/")

for file in files_to_copy:
    self.platform.comm.copy_from_host(f"{src / file}", f"{dst / file}")
```

I was a bit on the fence about whether to add this in benchkit natively
by extending the Benchmark class with for example a default function
"files_to_copy_before_build" that the user can override, or just provide
the copy tools and leave it up to the user. I ended up going with the
latter, but let me know if extending the Benchmark class is better.