perf-stat: json output, remote support and bug fix
This patch adds the support for the json output (using perf stat --json
option), add the support of remote for the PerfStatWrap wrapper (based
on the previous PRs) and fix a bug with french locales, i.e. it was
using "," instead of "." for the decimal delimiter, breaking the json
output. For that, we force the locale by setting an environment variable
in the PerfStatWrap wrapper (we added the mechanisms to update the
environment in command wrappers).

Signed-off-by: Antonio Paolillo <apaolill@gmail.com>