lwchart: fix 2 bugs
Fix a first bug with race condition on the file system when saving the
charts. Since the actual save did not completely complete before the
next "filename allocation", two consecutive charts generation can get
the same filename, ending up with the second chart overriding the first
one. We had the creation of an empty file just after allocation to
minimize the risk it happens.

Fix a second bug with lwchart crashing when the dependencies are missing
when calling the function `generate_chart_from_single_csv`. Add the
missing check to avoid that bug.

Signed-off-by: Antonio Paolillo <apaolill@gmail.com>