Allow pinning CPUs using TaksetWrap
Add option of `set_all_cpus` which sets the CPU affinity for all threads
running. This requires that the benchmark being used passes `nb_threads` to the
command_wrapper.

If `set_all_cpus` is not set or `nb_threads` is not given, then keeps previous
behavior -- that is, pinning only the master thread.

Signed-off-by: Rafael Chehab <chehabrafael@gmail.com>