Allow SIGTERM kill of remote aync calls
Previous implementation kills the `ssh -t <remote> <async>` command, which will
generate a soft SIGTERM kill signal in the local process but a stronger signal
on the remote asynchronous call.

For benchmarks/calls with a cleanup "graceful" exit, a SIGTERM signal should be
sent directly for the asynchronous call.

This benchmark creates an individual `sshd` link only for the asynchronous call,
which is then used to find the remote pid and kill it -- thus guaranteeing a
graceful exit.

Signed-off-by: Rafael Chehab <chehabrafael@gmail.com>