[WIP] Increase DB size for readreverse/sequential
Distinct from other read benchmarks, the readreverse and readsequential will
read all elements in the DB only once and then stop.

For the default size of the DB created, this leads to a very short duration and
as such the size of the DB should be increased for these benchmarks.

Signed-off-by: Rafael Chehab <chehabrafael@gmail.com>