feat(packagemanagers): Add default package manager
Add a default package manager when no known package manager can be
found. This package manager will always return `true` when asked if a
certain package is installed.

This is to allow benchkit to run on platform for which it doesn't know
the package manager, even if it can't give info about installed
packages.