refactor(benchmark): Return default value from `get_tilt_var_names`
Using `tilt` is deprecated, so the `get_tilt_var_names` function should
not longer be used, but since, every time it was called, this function
returned a `notImplementedError`, this function still had to be
overwritten every time you wanted to create a new benchmark class.
Returning an empty list by default fixes this.