feat(campaign): Add ability to symlink latest campaign
When the argument `symlink_latest` is set to `True` for a given
campaign, it will create a symlink either to the results folder or the
resulting `CSV` file.
These symlinks use the same naming conventions as the other files, but
use `_latest` as a suffix instead of using the current time.

This is done to allow people or tools that are inspecting the results
to easily get the latest result, without removing the older results.