To make an apple pie you must first...
... invent the universe.

This patch introduces a more formal system to represent commands and
their arguments.
Eventually, it is meant to replace the `shell_out` function and its
variants, that are all based on text to represent commands, whereas we
aim at a representation approaching an AST (or IR) for shell commands,
which would support concatenating, executing in parallel, recursion
(wrappers of commands that contain commands) and, more generally,
composition of command constructs.

This implementation is not mature and must be seen as a draft for
further discussing the roadmapped implementation.

Signed-off-by: Antonio Paolillo <apaolill@gmail.com>