Introduce CLI interface (#218)
To improve the "first impression" of benchkit, we introduce a CLI
interface that helps bootstrap benchmarking projects. The CLI provides:

- Benchmark template generation with multiple layout options
- Installation of the benchkit virtual environment and repository under
~/.benchkit
- Activation helpers and shell-aware suggestions for environment setup

Eventually, this enables a streamlined install process requiring only:

  pip3 install --user pybenchkit
  ~/.local/bin/benchkit install
  . ~/.benchkit/activate

With that, the host machine is ready to run benchmarks. Benchmark
skeletons can then be generated and run as follows:

  benchkit init --command="echo bench" --single-file --git
  benchkit run

All that's left is to profit from the benchmark results.

Signed-off-by: Antonio Paolillo <apaolill@gmail.com>