feat(shellasync): Option to ignore certain error codes
Add the option to allow certain error codes that are produced by an
asynchronous shell.
This allows you to spawn an asynchronous command that you expect to
give an error, but still let it continue normally.

This is already implemented for synchronous shells, and the
implementation for asynchronous shells is largely based on this
implementation.