lscpu: handle socket(s) without integer value (#263)
This patch improves the sockets() method in lscpu.py for platforms that
have no amount of sockets specified in the lscpu utility.

Previously, accessing the `Socket(s)` property could throw an exception
if the value was missing (e.g., `Socket(s): -` on Raspberry Pi 5).
We now log a warning and returns a default value of 1 to prevent
crashes.

Signed-off-by: Vic De Hondt <vicdehondt@users.noreply.github.com>