Added a signal attachment (#267)
This attachment can be used to send signals to the attached benchmark.

A possible use case for this attachment is to use it multiple times in
the list of attachments. The first attachment could be a signal that
stops the running process (SIGSTOP), and the last one a signal that
tells it to continue (SIGCONT). This enables attachments to have the
time to attach to the process before the benchmark continues.

We added an example of such usage in the speedup stack command wrapper.

Signed-off-by: theodegeest <theo.degeest@gmail.com>