klockstat must wait eBPF program is attached (#268)
These changes ensure that the klockstat attachment has attached its eBPF
as expected. In case the eBPF is not attached after a few seconds, a
timeout error is raised.

After these changes, the output from the klockstat attachment is more
reliable because it can monitor the start of the benchmark. This also
makes it possible to use this attachment on small benchmarks.

Signed-off-by: theodegeest <theo.degeest@gmail.com>