Implement offcputime attachment (#269)
These changes add a new attachment for the offcputime libbpf-tools
example. This eBPF-based attachment records the total time each head is
not actively scheduled on a CPU core.

A possible use for this attachment is detecting load imbalance in
multithreaded benchmarks. Such a use case can be found in the
speedupstack wrapper.

Signed-off-by: theodegeest <theo.degeest@gmail.com>